+++ Tests/CSharp/TriviaTests.cs
@@ -1,4 +1,7 @@
-﻿using System.Threading.Tasks;
+﻿using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using ICSharpCode.CodeConverter.Common;
+using ICSharpCode.CodeConverter.CSharp;
 using ICSharpCode.CodeConverter.Tests.TestRunners;
 using Xunit;
 
@@ -346,5 +349,27 @@ internal partial class DisplayAttribute : Attribute
 {
     public string Name { get; set; }
 }");
+    }
+
+    [Fact]
+    public async Task TestCarryingOverErrorTriviaAddedByConverterAsync()
+    {
+        var vbCode = @"Public Class VisualBasicClass
+    Public Class TestClass
+        Public Property A As Integer
+        Public Property B As integer
+    End class
+   Public Sub New()
+        Dim y As  New TestClass() With { .A = 1, .B = .A 
+        }
+   End Sub
+End Class";
+
+        var options = new TextConversionOptions(DefaultReferences.NetStandard2);
+        var conversionResult = await ProjectConversion.ConvertTextAsync<VBToCSConversion>(vbCode, options);
+
+        var regex = new Regex(@"#error Cannot convert \w+ - see comment for details\s+ \/\* Cannot convert.*?\*\/", RegexOptions.Singleline);
+        Assert.Equal(1, conversionResult.Exceptions.Count);
+        Assert.Matches(regex, conversionResult.ConvertedCode);
     }
 }